+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ### Added
 
+* added `stickyHeader` to disable/enable timeline header sticking on scroll.
 * removed `fullUpdate` prop and functionality. Labels rely on `position: sticky` to show for items that start before `visibleTimeStart`. This (should) greatly improve scroll performance.
 * removed extraneous css such as `text-align: center` on `.rct-item`, `.rct-item-overflow` to simplify the dom structure of `Item.js`